requirements.txt
fastapi
uvicorn
pydantic
openai
python-multipart
aiofiles
jinja2
python-dotenv